+++ src/pipi/PiPiEditor.h
@@ -24,9 +24,9 @@ namespace PiPi {
             PiPiEditor* addField(std::string fieldName, PiPiFieldType type, unsigned int pageIndex, double x, double y, double width, double height);
 			
             PiPiEditor* removeField(std::string fieldName);
-            PiPiEditor* removeField(std::string fieldName, unsigned int pageIndex);
-            PiPiEditor* removeField(std::string fieldName, unsigned int pageIndex, double x, double y);
-            PiPiEditor* removeField(std::string fieldName, unsigned int pageIndex, double x, double y, double width, double height);
+            PiPiEditor* removeField(std::string fieldName, long pageIndex);
+            PiPiEditor* removeField(std::string fieldName, long pageIndex, double x, double y);
+            PiPiEditor* removeField(std::string fieldName, long pageIndex, double x, double y, double width, double height);
         
 			PiPiEditor* renameField(std::string oldFieldName, std::string newFieldName);
